More refactoring. NOTE: untested code
