+++ app/[locale]/summary/Summary.tsx
@@ -3,55 +3,19 @@
 import Link from "next/link";
 import { useRouter } from 'next/navigation';
 import { useContext, useState, useEffect } from "react";
-import useSWR from "swr";
 import GiftCardComponent from '@/components/GiftCard';
 import Order from "@/components/Order";
+import ContactComponent from "./Contact";
 import { AppContext } from "@/context/AppContext";
-import { fetchAPI } from "@/lib/api";
-import { getContactForm } from '@/utils/helpers';
-import { useTranslation } from "@/context/useTranslation";
-import { ContactForm, Customer, Item } from "@/utils/models";
-
-type ContactProps = {
-  customer: Customer;
-  items: Item[];
-  locale: string;
-}
-
-const ContactComponent = ({customer, items, locale}: ContactProps) => {
-  const { translation } = useTranslation(locale);
-  const { data } = useSWR<ContactForm[]>(['/contact-forms',locale], url => fetchAPI(url,{},{
-    locale,
-    populate: ['contactForm','itemTypes']
-  }));
-  const fields = data && getContactForm(data,items);
-
-  if(!fields) return null;
-  return (
-    <div className='mb-5 text-primary-900 dark:text-primary-100'>
-      <div className='shadow-lg rounded p-4 flex flex-col gap-2'>
-        {fields.map(row =>
-        <div
-        className='flex flex-col sm:flex-row'
-        key={row.fieldName}>
-          <div className='flex-1 text-sm sm:text-base'>{row.label}</div>
-          <div className='flex-1'>{
-          row.type === 'checkbox' ?
-          customer.attributes[row.fieldName] ? translation.yes : translation.no
-          :customer.attributes[row.fieldName]
-          }</div>
-        </div>)}
-    </div>
-    </div>
-  );
-}
+import { ContactForm } from "@/utils/models";
 
 type SummaryProps = {
-  locale: string
+  locale: string;
+  translation: Record<string,string>;
+  contactForms: ContactForm[];
 }
 
-const Summary = ({locale}: SummaryProps) => {
-  const { translation } = useTranslation(locale);
+const Summary = ({locale, translation, contactForms}: SummaryProps) => {
   const {customer, items, isEmpty, order} = useContext(AppContext);
   const router = useRouter();
   const [ termsAccepted, setTermsAccepted ] = useState(false);
@@ -63,7 +27,12 @@ const Summary = ({locale}: SummaryProps) => {
   return (
     <div className='container mx-auto max-w-3xl py-4'>
       <div className='bg-secondary-50 dark:bg-secondary-800 rounded'>
-        <ContactComponent items={items} customer={customer} locale={locale}/>
+        <ContactComponent
+          items={items}
+          customer={customer}
+          translation={translation}
+          contactForms={contactForms}
+          />
       </div>
       <div className='bg-secondary-50 dark:bg-secondary-800 rounded shadow-lg p-4'>
         <Order items={items} locale={locale} translation={translation}></Order>
