+++ app/[locale]/summary/page.tsx
@@ -1,13 +1,29 @@
-import Component from './Component';
-
-
+import { fetchAPI } from '@/lib/api';
+import Summary from './Summary';
+import { ContactForm as ContactFormType } from "@/utils/models";
+import { getTranslation } from '@/utils/translationHelper';
 type Props = {
   params: {
     locale: string
   }
 }
-const ContactPage = ({params: {locale}}: Props) => {
-  return <Component locale={locale}/>
+
+const getContactForms = async (locale: string) => {
+  try {
+    const data = await fetchAPI<ContactFormType[]>('/contact-forms', { cache: 'no-store' },{
+      locale,
+      populate: ['contactForm','itemTypes']
+    });
+    return data;
+  } catch(error) {
+    return [];
+  }
+}
+
+const ContactPage = async ({params: {locale}}: Props) => {
+  const contactForms = await getContactForms(locale);
+  const translation = await getTranslation(locale);
+  return <Summary locale={locale} contactForms={contactForms} translation={translation}/>
 }
 
 export default ContactPage;
